--- packages/inspector-wpt/src/test.ts
@@ -1,407 +1,0 @@
-// interface a {
-//   numSteps: 1;
-//   run: 1;
-//   tester: string;
-//   minify_total: -1;
-//   cpuTimes: {
-//     Paint: 0;
-//     Idle: 11064;
-//   };
-//   responses_200: 111;
-//   testStartOffset: 0;
-//   bytesOut: 24019;
-//   gzip_savings: 20768;
-//   requestsFull: 114;
-//   start_epoch: 0;
-//   connections: 43;
-//   base_page_cdn: 'Akamai';
-//   bytesOutDoc: 17223;
-//   result: 0;
-//   final_base_page_request_id: 'CF935D531E511CDCF5B81EF9492ADB06';
-//   score_cookies: -1;
-//   basePageSSLTime: 141;
-//   docTime: 8062;
-//   domContentLoadedEventEnd: 5254;
-//   image_savings: 4712;
-//   requestsDoc: 92;
-//   firstMeaningfulPaint: 4756;
-//   firstTextPaint: 4007;
-//   firstPaint: 3830.0249999993;
-//   score_cdn: 96;
-//   'cpu.Idle': 11064;
-//   optimization_checked: 1;
-//   image_total: 231748;
-//   score_minify: -1;
-//   gzip_total: 1539047;
-//   responses_404: 0;
-//   loadTime: 8062;
-//   URL: 'https://www.gazzetta.it';
-//   score_combine: -1;
-//   firstContentfulPaint: 4007;
-//   firstLayout: 3786;
-//   score_etags: -1;
-//   loadEventStart: 8061;
-//   cpuTimesDoc: {
-//     Paint: 0;
-//     Idle: 8062;
-//   };
-//   minify_savings: -1;
-//   score_progressive_jpeg: 0;
-//   domInteractive: 5108;
-//   score_gzip: 99;
-//   'cpu.Paint': 0;
-//   score_compress: 98;
-//   domContentLoadedEventStart: 5109;
-//   final_url: 'https://www.gazzetta.it/';
-//   bytesInDoc: 1445636;
-//   firstImagePaint: 4115;
-//   'score_keep-alive': 100;
-//   loadEventEnd: 8097;
-//   cached: 0;
-//   score_cache: 50;
-//   responses_other: 3;
-//   main_frame: 'F39498E68E5D4A5A8CD7E132415E2C2A';
-//   fullyLoaded: 11064;
-//   requests: [];
-//   final_base_page_request: 0;
-//   TTFB: 1289;
-//   bytesIn: 1605773;
-//   test_run_time_ms: 16269;
-//   browser_version: '72.0.3626.96';
-//   base_page_dns_server: 'dnsgeo5.tuonomeregistrar.com';
-//   fullyLoadedCPUms: 8019;
-//   'PerformancePaintTiming.first-contentful-paint': 4006.7199999994;
-//   base_page_ip_ptr: 'a104-113-4-166.deploy.static.akamaitechnologies.com';
-//   eventName: 'Step_1';
-//   detected: {
-//     'Tag Managers': 'Google Tag Manager';
-//     Analytics: 'Chartbeat,Dynatrace,Google Analytics,SiteCatalyst';
-//     'Web Frameworks': 'ZURB Foundation 4.3.2';
-//     'JavaScript Libraries': 'Modernizr 2.6.2,jQuery 1.10.2';
-//     'JavaScript Frameworks': 'ExtJS';
-//     'Advertising Networks': 'Open AdStream';
-//     Miscellaneous: 'webpack';
-//   };
-//   base_page_cname: 'www2.gazzetta.it.edgekey.net';
-//   document_URL: 'https://www.gazzetta.it/';
-//   date: 1550156316.7607;
-//   'PerformancePaintTiming.first-paint': 3830.0249999993;
-//   domElements: 3590;
-//   browserVersion: '72.0.3626.96';
-//   document_origin: 'https://www.gazzetta.it';
-//   browser_name: 'Chrome';
-//   detected_apps: {
-//     jQuery: '1.10.2';
-//     'ZURB Foundation': '4.3.2';
-//     SiteCatalyst: '';
-//     Modernizr: '2.6.2';
-//     'Google Tag Manager': '';
-//     Chartbeat: '';
-//     Dynatrace: '';
-//     'Google Analytics': '';
-//     'Open AdStream': '';
-//     webpack: '';
-//     ExtJS: '';
-//   };
-//   fullyLoadedCPUpct: 10.014985015;
-//   domComplete: 8060;
-//   document_hostname: 'www.gazzetta.it';
-//   interactivePeriods: [[0, 4287]];
-//   SpeedIndex: 4458;
-//   visualComplete85: 4900;
-//   visualComplete90: 5000;
-//   visualComplete95: 5300;
-//   visualComplete99: 5300;
-//   visualComplete: 8500;
-//   render: 4000;
-//   lastVisualChange: 8500;
-//   chromeUserTiming: [
-//     {
-//       name: 'fetchStart';
-//       time: 15;
-//     },
-//     {
-//       name: 'unloadEventStart';
-//       time: 1295;
-//     },
-//     {
-//       name: 'unloadEventEnd';
-//       time: 1295;
-//     },
-//     {
-//       name: 'domLoading';
-//       time: 1296;
-//     },
-//     {
-//       name: 'responseEnd';
-//       time: 1425;
-//     },
-//     {
-//       name: 'firstLayout';
-//       time: 3785;
-//     },
-//     {
-//       name: 'firstPaint';
-//       time: 3830;
-//     },
-//     {
-//       name: 'firstMeaningfulPaintCandidate';
-//       time: 3830;
-//     },
-//     {
-//       name: 'firstMeaningfulPaintCandidate';
-//       time: 4006;
-//     },
-//     {
-//       name: 'firstTextPaint';
-//       time: 4006;
-//     },
-//     {
-//       name: 'firstContentfulPaint';
-//       time: 4006;
-//     },
-//     {
-//       name: 'firstImagePaint';
-//       time: 4115;
-//     },
-//     {
-//       name: 'firstMeaningfulPaint';
-//       time: 4755;
-//     },
-//     {
-//       name: 'firstMeaningfulPaint';
-//       time: 4755;
-//     },
-//     {
-//       name: 'firstMeaningfulPaintCandidate';
-//       time: 4755;
-//     },
-//     {
-//       name: 'domInteractive';
-//       time: 5108;
-//     },
-//     {
-//       name: 'domContentLoadedEventStart';
-//       time: 5109;
-//     },
-//     {
-//       name: 'domContentLoadedEventEnd';
-//       time: 5254;
-//     },
-//     {
-//       name: 'domComplete';
-//       time: 8060;
-//     },
-//     {
-//       name: 'loadEventStart';
-//       time: 8061;
-//     },
-//     {
-//       name: 'loadEventEnd';
-//       time: 8097;
-//     }
-//   ];
-//   'chromeUserTiming.fetchStart': 15;
-//   'chromeUserTiming.unloadEventStart': 1295;
-//   'chromeUserTiming.unloadEventEnd': 1295;
-//   'chromeUserTiming.domLoading': 1296;
-//   'chromeUserTiming.responseEnd': 1425;
-//   'chromeUserTiming.firstLayout': 3785;
-//   'chromeUserTiming.firstPaint': 3830;
-//   'chromeUserTiming.firstMeaningfulPaintCandidate': 4755;
-//   'chromeUserTiming.firstTextPaint': 4006;
-//   'chromeUserTiming.firstContentfulPaint': 4006;
-//   'chromeUserTiming.firstImagePaint': 4115;
-//   'chromeUserTiming.firstMeaningfulPaint': 4755;
-//   'chromeUserTiming.domInteractive': 5108;
-//   'chromeUserTiming.domContentLoadedEventStart': 5109;
-//   'chromeUserTiming.domContentLoadedEventEnd': 5254;
-//   'chromeUserTiming.domComplete': 8060;
-//   'chromeUserTiming.loadEventStart': 8061;
-//   'chromeUserTiming.loadEventEnd': 8097;
-//   blinkFeatureFirstUsed: {
-//     AnimatedCSSFeatures: [];
-//     CSSFeatures: {};
-//     Features: {};
-//   };
-//   TTIMeasurementEnd: 4287;
-//   LastInteractive: 4006;
-//   step: 1;
-//   effectiveBps: 164273;
-//   effectiveBpsDoc: 213441;
-//   domTime: 0;
-//   aft: 0;
-//   titleTime: 0;
-//   domLoading: 0;
-//   server_rtt: 0;
-//   smallImageCount: 23;
-//   bigImageCount: 0;
-//   maybeCaptcha: 0;
-//   pages: {
-//     details: 'http://localhost:30005/details.php?test=190214_EM_S&run=1';
-//     checklist: 'http://localhost:30005/performance_optimization.php?test=190214_EM_S&run=1';
-//     breakdown: 'http://localhost:30005/breakdown.php?test=190214_EM_S&run=1';
-//     domains: 'http://localhost:30005/domains.php?test=190214_EM_S&run=1';
-//     screenShot: 'http://localhost:30005/screen_shot.php?test=190214_EM_S&run=1';
-//   };
-//   thumbnails: {
-//     waterfall: 'http://localhost:30005/result/190214_EM_S/1_waterfall_thumb.png';
-//     checklist: 'http://localhost:30005/result/190214_EM_S/1_optimization_thumb.png';
-//     screenShot: 'http://localhost:30005/result/190214_EM_S/1_screen_thumb.png';
-//   };
-//   images: {
-//     waterfall: 'http://localhost:30005/results/19/02/14/EM/S/1_waterfall.png';
-//     connectionView: 'http://localhost:30005/results/19/02/14/EM/S/1_connection.png';
-//     checklist: 'http://localhost:30005/results/19/02/14/EM/S/1_optimization.png';
-//     screenShot: 'http://localhost:30005/getfile.php?test=190214_EM_S&file=1_screen.jpg';
-//   };
-//   rawData: {
-//     headers: 'http://localhost:30005/results/19/02/14/EM/S/1_report.txt';
-//     pageData: 'http://localhost:30005/results/19/02/14/EM/S/1_IEWPG.txt';
-//     requestsData: 'http://localhost:30005/results/19/02/14/EM/S/1_IEWTR.txt';
-//     utilization: 'http://localhost:30005/results/19/02/14/EM/S/1_progress.csv';
-//     trace: 'http://localhost:30005/getgzip.php?test=190214_EM_S&compressed=1&file=1_trace.json.gz';
-//   };
-//   domains: {
-//     'cdn-gl.imrworldwide.com': {
-//       bytes: 200561;
-//       requests: 5;
-//       cdn_provider: 'Amazon CloudFront';
-//       connections: 2;
-//     };
-//     'secure-it.imrworldwide.com': {
-//       bytes: 3683;
-//       requests: 4;
-//       connections: 1;
-//     };
-//     'www.facebook.com': {
-//       bytes: 72;
-//       requests: 1;
-//       connections: 1;
-//     };
-//     'www.googletagmanager.com': {
-//       bytes: 23685;
-//       requests: 1;
-//       cdn_provider: 'Google';
-//       connections: 1;
-//     };
-//     'www.google-analytics.com': {
-//       bytes: 17543;
-//       requests: 1;
-//       cdn_provider: 'Google';
-//       connections: 1;
-//     };
-//     'surveygizmobeacon.s3.amazonaws.com': {
-//       bytes: 69;
-//       requests: 1;
-//       connections: 1;
-//     };
-//     'mab.chartbeat.com': {
-//       bytes: 125;
-//       requests: 1;
-//       cdn_provider: 'Fastly';
-//       connections: 1;
-//     };
-//     'static.chartbeat.com': {
-//       bytes: 7854;
-//       requests: 1;
-//       cdn_provider: 'Amazon CloudFront';
-//       connections: 1;
-//     };
-//     'stadvtools.akamaized.net': {
-//       bytes: 143726;
-//       requests: 17;
-//       cdn_provider: 'Akamai';
-//       connections: 3;
-//     };
-//     'cdn.krxd.net': {
-//       bytes: 88366;
-//       requests: 4;
-//       cdn_provider: 'Fastly';
-//       connections: 1;
-//     };
-//     'beacon.krxd.net': {
-//       bytes: 76;
-//       requests: 1;
-//       connections: 1;
-//     };
-//     'consumer.krxd.net': {
-//       bytes: 189;
-//       requests: 1;
-//       cdn_provider: 'Fastly';
-//       connections: 1;
-//     };
-//     'd2bnxibecyz4h5.cloudfront.net': {
-//       bytes: 27190;
-//       requests: 1;
-//       cdn_provider: 'Amazon CloudFront';
-//       connections: 1;
-//     };
-//     'rum-dytrc.gazzetta.it': {
-//       bytes: 35275;
-//       requests: 2;
-//       connections: 2;
-//     };
-//     'api2-mtc.gazzetta.it': {
-//       bytes: 2692;
-//       requests: 2;
-//       cdn_provider: 'Akamai';
-//       connections: 2;
-//     };
-//     'www.gazzetta.it': {
-//       bytes: 52080;
-//       requests: 4;
-//       cdn_provider: 'Akamai';
-//       connections: 1;
-//     };
-//     'apicommunity.gazzetta.it': {
-//       bytes: 2436;
-//       requests: 1;
-//       connections: 1;
-//     };
-//     'dmpmetrics.rcsmetrics.it': {
-//       bytes: 339;
-//       requests: 1;
-//       connections: 1;
-//     };
-//     'smetrics.rcsmetrics.it': {
-//       bytes: 43;
-//       requests: 2;
-//       connections: 1;
-//     };
-//     'images2.gazzettaobjects.it': {
-//       bytes: 255926;
-//       requests: 32;
-//       cdn_provider: 'Akamai';
-//       connections: 6;
-//     };
-//     'js2.gazzettaobjects.it': {
-//       bytes: 188781;
-//       requests: 16;
-//       cdn_provider: 'Akamai';
-//       connections: 4;
-//     };
-//     'css2.gazzettaobjects.it': {
-//       bytes: 386392;
-//       requests: 10;
-//       cdn_provider: 'Akamai';
-//       connections: 8;
-//     };
-//     'components2.gazzettaobjects.it': {
-//       bytes: 147492;
-//       requests: 4;
-//       cdn_provider: 'Akamai';
-//       connections: 2;
-//     };
-//   };
-//   breakdown: {
-//     html: IbreakdownItem;
-//     js: IbreakdownItem;
-//     css: IbreakdownItem;
-//     image: IbreakdownItem;
-//     flash: IbreakdownItem;
-//     font: IbreakdownItem;
-//     video: IbreakdownItem;
-//     other: IbreakdownItem;
-//   };
-// }
