+++ packages/inspector-wpt/src/utils.ts
@@ -44,10 +44,16 @@ const loopUntillGetRespose = async url => {
   });
 };
 
-const audit = async (url: string, mobile: string, fvonly: string): Promise<any> => {
+const audit = async (
+  url: string,
+  mobile: string,
+  fvonly: string,
+  uastring?: string
+): Promise<any> => {
   try {
     console.log(`Getting data for ${url}`);
-    const runTestUrl = `${WPT_SERVER_URL}/runtest.php?f=json&url=${url}&mobile=${mobile}&fvonly=${fvonly}`;
+    const agentstring = uastring && `&uastring=${uastring}`;
+    const runTestUrl = `${WPT_SERVER_URL}/runtest.php?f=json&url=${url}&mobile=${mobile}&fvonly=${fvonly}${agentstring}`;
     const runTestResp = await fetch(runTestUrl);
     const runTestRespJSON = (await runTestResp.json()) as IRunTestRespose;
     const audit = await loopUntillGetRespose(runTestRespJSON.data.jsonUrl);
